requirements.txt
flask==2.2.5
streamlit==1.28.0
tensorflow==2.20.0
keras==2.12.0
numpy==1.24.3
opencv-python==4.7.0.72
pillow==9.5.0
mtcnn==0.1.0
scikit-learn==1.2.2
matplotlib==3.7.1
h5py==3.8.0
gunicorn==20.1.0
